Fix conflicts with PR #29
