+++ libentitlement_test.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/opencontainers/runtime-spec/specs-go"
-	"github.com/stretchr/testify/require"
-
+	"github.com/docker/libentitlement/defaults"
 	"github.com/docker/libentitlement/entitlement"
 	secprofile "github.com/docker/libentitlement/secprofile"
+	"github.com/opencontainers/runtime-spec/specs-go"
+	"github.com/stretchr/testify/require"
 )
 
 func testSpec() *specs.Spec {
@@ -42,17 +42,18 @@ func testSpec() *specs.Spec {
 
 func TestRegisterDummyEntitlement(t *testing.T) {
 	spec := testSpec()
-	profile := secprofile.NewProfile(spec)
+	profile := secprofile.NewOCIProfile(spec)
 
 	entMgr := NewEntitlementsManager(profile)
 
 	// Add a dummy "foo.bar.cap-sys-admin" void entitlement that adds CAP_SYS_ADMIN
-	capSysAdminVoidEntCallback := func(profile *secprofile.Profile) (*secprofile.Profile, error) {
-		if profile == nil {
-			return nil, fmt.Errorf("CapSysAdminVoidEntCallback - profile is nil")
+	capSysAdminVoidEntCallback := func(profile secprofile.Profile) (secprofile.Profile, error) {
+		ociProfile, ok := profile.(*secprofile.OCIProfile)
+		if !ok {
+			return nil, fmt.Errorf("%s: error converting to OCI profile", defaults.NetworkNoneEntFullID)
 		}
 
-		profile.AddCaps("CAP_SYS_ADMIN")
+		ociProfile.AddCaps("CAP_SYS_ADMIN")
 
 		return profile, nil
 	}
